requirements.txt
fastapi 
uvicorn 
sqlalchemy 
pyodbc 
python-jose[cryptography] 
passlib[bcrypt] 
python-multipart
python-dotenv
pyjwt
tzdata
jinja2

PyPDF2
langchain
langchain-community
python-docx
requests
langchain-openai
docx2txt
azure-search-documents
azure-identity
python-multipart
pypdf